refine registry to support quantization and construct llama2 mode;

+++ test/testOperation/testAdd.cpp
@@ -1,5 +1,6 @@
 #include <gtest/gtest.h>
 #include <cstdlib>
+#include <cstdint>
 
 #include <cuda_runtime_api.h>
 
@@ -13,22 +14,22 @@ TEST(test_add, gpu_add) {
     using eUTIL::Tensor;
 
 
-    Tensor<int> cpuA(DeviceType::kCpu, kCount);
-    Tensor<int> cpuB(DeviceType::kCpu, kCount);
-    Tensor<int> cpuO(DeviceType::kCpu, kCount);
+    Tensor<float> cpuA(DeviceType::kCpu, kCount);
+    Tensor<float> cpuB(DeviceType::kCpu, kCount);
+    Tensor<float> cpuO(DeviceType::kCpu, kCount);
 
     // Initialize
     for (int i = 0; i < kCount; ++i) {
-        cpuA[i] = std::rand() % 10;
-        cpuB[i] = std::rand() % 10;
+        cpuA[i] = static_cast<float>(std::rand() % 10);
+        cpuB[i] = static_cast<float>(std::rand() % 10);
     }
 
     auto addKernelCpu = eCEL::KernelFactory::getAddKernel();
     addKernelCpu(cpuA, cpuB, cpuO, nullptr);
 
-    Tensor<int> gpuA = cpuA.cuda();
-    Tensor<int> gpuB = cpuB.cuda();
-    Tensor<int> gpuO(DeviceType::kCuda, kCount);
+    Tensor<float> gpuA = cpuA.cuda();
+    Tensor<float> gpuB = cpuB.cuda();
+    Tensor<float> gpuO(DeviceType::kCuda, kCount);
 
     auto addKernelCuda = eCEL::KernelFactory::getAddKernel();
     addKernelCuda(gpuA, gpuB, gpuO, nullptr);
